sys,fw: allow to create multiprotocol rules
Now you can add rules to allow multiple protocols.
For example you can add a rule to allow dport/sport for both TCP
and UDP.

There're two options to allow a port:

Statement {
 Name: tcp
 Values:
   Key: dport
   Value: 1234
}

Statement {
 Name: meta
 Values:
   Key: l4proto
   Value: tcp,udp

   Key: dport
   Value: 1234
}

Closes #951.

+++ daemon/firewall/nftables/exprs/utils.go
@@ -130,7 +130,7 @@ func getICMPv6RejectCode(reason string) uint8 {
 // If the protocol is not in our list, we'll use the value as decimal.
 // So for example IPPROTO_ENCAP (0x62) must be specified as 98.
 // https://pkg.go.dev/golang.org/x/sys/unix#pkg-constants
-func getProtocolCode(value string) (uint32, error) {
+func getProtocolCode(value string) (byte, error) {
 	switch value {
 	case NFT_PROTO_TCP:
 		return unix.IPPROTO_TCP, nil
@@ -174,5 +174,5 @@ func getProtocolCode(value string) (uint32, error) {
 	if err != nil {
 		return 0, err
 	}
-	return uint32(prot), nil
+	return byte(prot), nil
 }
